+++ app/services/pdf_generator/cover_page_generator.rb
@@ -90,7 +90,14 @@ class PDFGenerator
 
         header(prawn_document, "Download Information")
         prawn_document.text "Date Rendered: #{Time.current.strftime('%Y-%m-%d %I:%M:%S %p %Z')}"
-        prawn_document.text "Available Online at: <u><a href='#{helper.solr_document_url(id: resource.id)}'>#{helper.solr_document_url(id: resource.id)}</a></u>", inline_format: true
+        resource_link = if resource.decorate.public_readable_state?
+                          identifier = Ark.new(resource.identifier.first)
+                          identifier.uri
+                        else
+                          IdentifierService.url_for(resource)
+                        end
+
+        prawn_document.text "Available Online at: <u><a href='#{resource_link}'>#{resource_link}</a></u>", inline_format: true
       end
     end
 
@@ -98,10 +105,6 @@ class PDFGenerator
 
       def holding_location_text(holding_location)
         ControlledVocabulary.for(:holding_location).find(holding_location).label
-      end
-
-      def helper
-        @helper ||= ManifestBuilder::ManifestHelper.new
       end
 
       def rights_statement_label(statement)